+++ dataclass_wizard/bases_meta.py
@@ -25,6 +25,10 @@ from .type_def import E
 from .utils.type_conv import date_to_timestamp, as_enum
 
 
+# global flag to determine if debug mode was ever enabled
+_debug_was_enabled = False
+
+
 def _as_enum_safe(cls: type, name: str, base_type: Type[E]) -> Optional[E]:
     """
     Attempt to return the value for class attribute :attr:`attr_name` as
@@ -99,9 +103,12 @@ class BaseJSONWizardMeta(AbstractMeta):
         cls_dumper = get_dumper(dataclass, create=create)
 
         if cls.debug_enabled:
+            global _debug_was_enabled
+            if not _debug_was_enabled:
+                _debug_was_enabled = True
+                LOG.setLevel('DEBUG')
+                LOG.info('DEBUG Mode is enabled')
 
-            LOG.setLevel('DEBUG')
-            LOG.info('DEBUG Mode is enabled')
             # Decorate all hooks so they format more helpful messages
             # on error.
             load_hooks = cls_loader.__LOAD_HOOKS__